feat: split component at runtime builder stage
This allows for the ability to register only a single function, and not an entire component. I don't like the naming of ComponentFunctionEx, so I may change that in the future.

I'm also slightly concerned that we're duplicating the data of each component 3-ways. I'm hoping that the shared-memory model of the JVM fixes this. Worst case scenario, it's only an extra ~2 bytes per ComponentFunctionEx.